# Speed up link checker by not loading historical API docs