Closes https://github.com/Qiskit/documentation/issues/461.

Before, the script took me about 42 seconds. Now, about 5.7:

```
❯ hyperfine 'npm run check:links'
Benchmark 1: npm run check:links
  Time (mean ± σ):      5.680 s ±  0.143 s    [User: 7.285 s, System: 0.471 s]
  Range (min … max):    5.548 s …  5.964 s    10 runs
 
```

This PR differentiates between files we want to check vs. files that we only need to know they exist because other files might link to them. For the latter, we do still need to read in those files to determine their anchors. 

The key insight of this PR is that 99% of the historical API docs can be ignored because the stable docs don't link to them. We only need to load a couple of historical API docs because our migration guides link to them. 

--

I think we probably will want to add an argument to allow checking links for historical API docs: https://github.com/Qiskit/documentation/issues/495. This should work well with that change; we'd update the list based off of the `Arguments` passed in.